+++ src/assets/tinymce/TinyMCEAsset.php
@@ -2,8 +2,10 @@
 
 namespace spicyweb\tinymce\assets\tinymce;
 
+use Craft;
 use craft\web\AssetBundle;
 use craft\web\assets\cp\CpAsset;
+use spicyweb\tinymce\Plugin;
 
 /**
  * Class FieldAsset
@@ -28,7 +30,31 @@ class TinyMCEAsset extends AssetBundle
             'tinymce.min.js',
         ];
 
+        $this->_loadTranslationFile(Plugin::$plugin->language->mapLanguage(Craft::$app->language));
+
         parent::init();
+    }
+
+    private function _loadTranslationFile(string $language): void
+    {
+        $languagesDir = Craft::$app->getPath()->getConfigPath() .
+            DIRECTORY_SEPARATOR . 'tinymce' . DIRECTORY_SEPARATOR . 'languages';
+
+        try {
+            if (is_dir($languagesDir)) {
+                $languageFile = $languagesDir . DIRECTORY_SEPARATOR . $language . '.js';
+
+                if (is_file($languageFile)) {
+                    $contents = stripslashes(file_get_contents($languageFile));
+
+                    if (preg_match('/tinymce.addI18n\(\'([a-zA-Z-_]+)\',(.+)\);/s', $contents, $matches)) {
+                        Craft::$app->getView()->registerJs($contents);
+                    }
+                }
+            }
+        } catch (\Exception $e) {
+            // Just do nothing
+        }
     }
 }