+++ src/fields/TinyMCE.php
@@ -206,7 +206,7 @@ class TinyMCE extends HtmlField
             'elementSiteId' => (string)$elementSite->id,
             'allSites' => $allSites,
             'direction' => $this->getOrientation($element),
-            'language' => $language,
+            'language' => Plugin::$plugin->language->mapLanguage($language),
             'translations' => $translations,
         ];
 
@@ -415,7 +415,6 @@ class TinyMCE extends HtmlField
 
     private function _loadTranslations(): array
     {
-        // TODO: would be good to load translation packages
         $messages = [
             'File',
             'New document',
@@ -544,7 +543,11 @@ class TinyMCE extends HtmlField
         $translations = [];
 
         foreach ($messages as $message) {
-            $translations[$message] = Craft::t('tinymce', $message);
+            $translatedMessage = Craft::t('tinymce', $message);
+
+            if ($message !== $translatedMessage) {
+                $translations[$message] = $translatedMessage;
+            }
         }
 
         return $translations;