+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Added the ability to load official TinyMCE translation files from the `config/tinymce/languages` directory
+- Added `spicyweb\tinymce\services\LanguageService`
+
 ## 1.2.1 - 2022-10-31
 
 ### Fixed